services/orchestrator/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
tenacity>=8.3.0
python-multipart>=0.0.9
groq
redis>=5.0.0
langgraph>=0.0.10
